mvm_gui/requirements.txt
PyQt5
PyQtGraph
PySerial
PyYaml